// app/src/interfaces/minikube.ts
// Generated by https://quicktype.io

export interface MinikubeVersion {
  commit?: string;
  minikubeVersion?: string;
}

export enum MinikubeStatus {
  Unknown = 'UNKNOWN',
  Running = 'RUNNING',
  Stopped = 'STOPPED',
  NotFound = 'NOTFOUND',
}

export interface MinikubeClusterStatus {
  Name: string;
  Host: string;
  Kubelet: string;
  APIServer: string;
  Kubeconfig: string;
  Worker: boolean;
  status: MinikubeStatus;
}